requirements.txt
amqp==5.1.1
asgiref==3.5.2
aspose-words==22.10.0
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.8
django-allauth==0.51.0
django-cors-headers==3.8.0
django-decouple==2.1
django-shortuuidfield==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docusign-esign==3.18.0
docx==0.2.4
exception==0.1.0
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
incremental==22.10.0
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
nose==1.3.7
oauthlib==3.2.1
pbr==5.10.0
Pillow==8.3.2
platformdirs==2.5.2
prompt-toolkit==3.0.31
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
zipp==3.8.1
zope.interface==5.5.0
